requirements.txt
pypdf
faiss-cpu
sentence-transformers
pandas
pyarrow
tqdm
openai
python-dotenv
matplotlib
scikit-learn